segmentation

+++ checker/label_checker_inclusion_of_maps.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   label_checker_inclusion_of_maps.c                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ftothmur <ftothmur@student.42.fr>          +#+  +:+       +#+        */
+/*   By: cmissy <cmissy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/27 16:54:02 by ftothmur          #+#    #+#             */
-/*   Updated: 2020/02/28 17:04:34 by ftothmur         ###   ########.fr       */
+/*   Updated: 2020/02/28 18:21:56 by cmissy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int					label_checker_inclusion_of_maps(t_vector *label_ptr_keys, t_hash_map *map_of_label_words)
 {
-	size_t			i;
+	int				i;
 	void			*curr_label_name;
 
 	i = 0;
